update code for async WPJ
update code for async WPJ

+++ ADAuthenticationBroker/ADAuthenticationBroker/NSString+ADBrokerHelperMethods.m
@@ -414,7 +414,7 @@ static inline void Encode3bytesTo4bytes(char* output, int b0, int b1, int b2)
 {
     const char* str = [self UTF8String];
     unsigned char result[CC_SHA256_DIGEST_LENGTH];
-    CC_SHA256(str, strlen(str), result);
+    CC_SHA256(str, (int)strlen(str), result);
     
     NSMutableString *ret = [NSMutableString stringWithCapacity:CC_SHA256_DIGEST_LENGTH*2];
     for(int i = 0; i<CC_SHA256_DIGEST_LENGTH; i++)
